add type checking
